+++ src/main/java/online/bottler/label/application/LabelService.java
@@ -2,15 +2,13 @@ package online.bottler.label.application;
 
 import java.util.List;
 import lombok.RequiredArgsConstructor;
-import online.bottler.global.exception.ApplicationException;
 import online.bottler.label.application.port.out.LabelPersistencePort;
 import online.bottler.label.domain.Label;
-import online.bottler.label.domain.LabelType;
 import online.bottler.label.application.command.LabelCommand;
 import online.bottler.label.application.port.in.LabelUseCase;
 import online.bottler.label.application.response.LabelResponse;
+import online.bottler.label.domain.LabelType;
 import online.bottler.scheduler.LabelScheduler;
-import online.bottler.user.application.port.in.UserUseCase;
 import online.bottler.user.domain.User;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
@@ -19,7 +17,6 @@ import org.springframework.transaction.annotation.Transactional;
 @RequiredArgsConstructor
 public class LabelService implements LabelUseCase {
     private final LabelPersistencePort labelPersistencePort;
-    private final UserUseCase userUseCase;
     private final LabelScheduler labelScheduler;
 
     @Transactional
@@ -39,22 +36,12 @@ public class LabelService implements LabelUseCase {
         return labels.stream().map(Label::toLabelResponse).toList();
     }
 
-    @Transactional
-    public LabelResponse createFirstComeFirstServedLabel(Long userId) {
-        User user = userUseCase.findById(userId);
+    public List<Label> findByLabelType(LabelType labelType) {
+        return labelPersistencePort.findByLabelType(labelType);
+    }
 
-        List<Label> firstComeLabels = labelPersistencePort.findByLabelType(LabelType.FIRST_COME);
-
-        for (Label label : firstComeLabels) {
-            boolean hasLabel = labelPersistencePort.existsUserLabelByUserAndLabel(user, label);
-            if (!hasLabel && label.isOwnedCountValid()) {
-                labelPersistencePort.updateOwnedCount(label);
-                labelPersistencePort.createUserLabel(user, label);
-                return label.toLabelResponse();
-            }
-        }
-
-        throw new ApplicationException("모든 선착순 뽑기 라벨이 마감되었습니다.");
+    public boolean isLabelExistsByUserAndLabel(User user, Label label) {
+        return labelPersistencePort.existsUserLabelByUserAndLabel(user, label);
     }
 
     @Transactional
