+++ src/main/java/online/bottler/label/application/port/in/LabelUseCase.java
@@ -4,6 +4,7 @@ import java.util.List;
 import online.bottler.label.application.command.LabelCommand;
 import online.bottler.label.application.response.LabelResponse;
 import online.bottler.label.domain.Label;
+import online.bottler.label.domain.LabelType;
 import online.bottler.user.domain.User;
 
 public interface LabelUseCase {
@@ -13,7 +14,9 @@ public interface LabelUseCase {
 
     List<LabelResponse> findUserLabels(Long userId);
 
-    LabelResponse createFirstComeFirstServedLabel(Long userId);
+    List<Label> findByLabelType(LabelType labelType);
+
+    boolean isLabelExistsByUserAndLabel(User user, Label label);
 
     List<LabelResponse> findFirstComeLabels();
